backend/requirements.txt
# Python 3.13.x required (match production environment on Render)
#
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Data handling
pydantic==2.10.6
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.11  # Pre-built wheel for Python 3.13

# AI Integration
anthropic==0.43.0
google-generativeai==0.8.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Development
black==24.10.0
ruff==0.8.4
